# src/swagger.yaml
openapi: 3.0.3
info:
  title: THORChain Memoless Registration API
  version: 1.0.0
  description: |
    A headless API service for THORChain memoless transaction registration. 
    This API manages a hot RUNE wallet to register memos on behalf of users, 
    eliminating the need for users to hold RUNE or manage THORChain transactions directly.
    
    ## Features
    - **Immediate Response**: Returns registration ID instantly
    - **Background Processing**: Handles transaction confirmation asynchronously  
    - **Comprehensive Validation**: Amount, reference ID, and dust threshold validation
    - **Suggested Amount Calculation**: Smart amount calculation with reference ID embedding
    - **Multi-chain Support**: Bitcoin, Ethereum, BSC, and all THORChain supported assets
    
    
  contact:
    name: Tyler Bond
    email: tyler@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.thorchain-memoless.com
    description: Production server
...
paths:
...
  /api/v1/assets:
    get:
      tags: [Assets]
      summary: Get Valid Assets
      description: |
        Returns all assets available for memoless registration. 
        Implements memoless.md Steps 1-2:
        - Queries THORChain pools endpoint
        - Filters for Available status
        - Excludes THOR chain assets and tokens
        - Sorted by descending balance_rune
      responses:
        '200':
          description: List of available assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  assets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
        '500':
          description: Failed to fetch assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/assets/{asset}:
    get:
      tags: [Assets]
      summary: Get Asset Details
      description: Get detailed information about a specific asset
      parameters:
        - name: asset
          in: path
          required: true
          description: Asset identifier (e.g., BTC.BTC, ETH.ETH)
          schema:
            type: string
            example: "BTC.BTC"
      responses:
        '200':
          description: Asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '404':
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/register:
    post:
      tags: [Registration]
      summary: Register Memo
      description: |
        Register a memo for memoless transactions and immediately return complete memo details.
        
        **⚠️ IMPORTANT VALIDATION REQUIREMENTS:**
        
        1. **Asset Validation**: The `asset` parameter must be from the list returned by `GET /api/v1/assets`
        2. **Memo Format**: Any non-empty string - THORChain will validate the memo format during registration
        
        **Process Flow (Synchronous):**
        - Validates asset against available THORChain pools
        - Creates MsgDeposit with REFERENCE:{asset}:{memo} format
        - Broadcasts transaction to THORChain
        - Immediately checks `/thorchain/memo/{txid}` for memo details
        - Returns complete memo reference data including reference ID
        
        **💡 Workflow:**
        1. First call `GET /api/v1/assets` to get available assets
        2. Select an asset from the returned list
        3. Prepare your THORChain memo
        4. Call this endpoint - it will complete the entire registration process synchronously
        5. Response includes reference ID ready for amount validation and deposits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        '201':
          description: Registration submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidAssetFormat:
                  summary: Invalid Asset Format
                  value:
                    success: false
                    error:
                      code: "INVALID_ASSET_FORMAT"
                      message: "Asset must be in format CHAIN.SYMBOL (e.g., BTC.BTC, ETH.ETH)"
                unsupportedAsset:
                  summary: Unsupported Asset
                  value:
                    success: false
                    error:
                      code: "UNSUPPORTED_ASSET"
                      message: "Asset 'INVALID.ASSET' is not available for memoless registration"
                      details:
                        providedAsset: "INVALID.ASSET"
                        supportedAssets: ["BTC.BTC", "ETH.ETH", "GAIA.ATOM", "AVAX.AVAX", "BSC.BNB"]
                        totalSupportedAssets: 23
                        suggestion: "Use GET /api/v1/assets to see all available assets"
                emptyMemo:
                  summary: Empty Memo
                  value:
                    success: false
                    error:
                      code: "EMPTY_MEMO"
                      message: "Memo cannot be empty"
        '409':
          description: Memo already registered for this asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/register/{registrationId}:
    get:
      tags: [Registration]
      summary: Get Registration Status
      description: |
        Get the current status of a registration including all data needed for Steps 7-9:
        - Reference ID and validation data
        - Inbound address and deposit info
        - Usage statistics and expiry information
      parameters:
        - name: registrationId
          in: path
          required: true
          description: Registration ID returned from POST /register
          schema:
            type: string
            example: "reg_abc123"
      responses:
        '200':
          description: Registration status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  registration:
                    $ref: '#/components/schemas/RegistrationStatus'
        '404':
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/preflight:
    post:
      tags: [Validation]
      summary: Preflight Check
      description: |
        Performs validation checks before registration to ensure the request will succeed.
        
        **Validation Steps:**
        - Validates asset is supported
        - Checks reference ID format and availability  
        - Validates amount format and dust threshold
        - Verifies memo format compatibility
        
        **Use this endpoint to:**
        - Pre-validate parameters before calling /register
        - Check if a reference ID is still available
        - Validate amount formatting and thresholds
        - Ensure memo will be accepted by THORChain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                internal_api_id:
                  type: string
                  description: Optional registration ID to validate against
                  example: "reg_abc123"
                asset:
                  type: string
                  description: Asset identifier (CHAIN.SYMBOL format)
                  default: "ETH.ETH"
                  example: "BTC.BTC"
                reference:
                  type: string
                  description: Reference ID to validate (if checking existing registration)
                  example: "00123"
                amount:
                  type: string
                  description: Amount to validate
                  default: "0.001"
                  example: "0.00100123"
                inputType:
                  type: string
                  enum: [asset]
                  default: asset
                  description: Amount input type (only asset supported)
      responses:
        '200':
          description: Preflight check passed - proceed with transaction
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Preflight check passed - proceed with transaction"
                  data:
                    type: object
                    properties:
                      current_uses:
                        type: integer
                        description: Current number of times this memo has been used
                        example: 0
                      max_uses:
                        type: integer
                        description: Maximum number of times this memo can be used
                        example: 3
                      memo:
                        type: string
                        description: The actual registered memo from THORChain
                        example: "DONATE:BTC.BTC"
                      inbound_address:
                        type: string
                        description: Chain-specific inbound address for deposits
                        example: "bc1qalgy3f70we7p3ymp36c2njepccr90nrj5wr535"
                      qr_code:
                        type: string
                        description: QR code string in chain-specific format
                        example: "bitcoin:bc1qalgy3f70we7p3ymp36c2njepccr90nrj5wr535?amount=0.00100033"
                      qr_code_data_url:
                        type: string
                        description: Base64 encoded PNG image data URL
                        example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                      time_remaining:
                        type: string
                        description: Time until registration expires
                        example: "24h"
                      blocks_remaining:
                        type: integer
                        description: Number of blocks until registration expires
                        example: 14400
                      seconds_remaining:
                        type: integer
                        description: Approximate seconds until registration expires (blocks * 6)
                        example: 86400
        '400':
          description: Preflight check failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: "PREFLIGHT_FAILED"
                      message:
                        type: string
                        example: "Reference ID is not registered yet - available for registration"
                  data:
                    type: object
                    properties:
                      current_uses:
                        type: integer
                        example: 0
                      max_uses:
                        type: integer
                        example: 3
                      time_remaining:
                        type: string
                        example: "No expiry"
                      blocks_remaining:
                        type: integer
                        example: 0
                      seconds_remaining:
                        type: integer
                        example: 0
  /api/v1/suggest-amounts:
    post:
      tags: [Validation]
      summary: Get Amount Suggestions
      description: |
        Calculates suggested amounts for memoless transactions with both rounding up and rounding down options.
        
        **Use Cases:**
        - User wants to send exactly their available balance (round down)
        - User can send slightly more for optimal validation (round up)
        - Comparing both options before making a transaction
        
        **Algorithm:**
        1. Normalizes input amount to asset decimal precision
        2. Embeds reference ID in decimal places
        3. Provides two suggestions:
           - **Round Up**: Ensures suggested > requested (current behavior)
           - **Round Down**: Ensures suggested ≤ requested (new behavior)
        
        **Important:** Both suggestions will be valid for memoless protocol, but the round down option may be lower than the user's requested amount.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - asset
                - reference
                - requested_amount
              properties:
                asset:
                  type: string
                  description: Asset identifier (CHAIN.SYMBOL format) - decimals will be automatically fetched from THORChain pools
                  example: "BTC.BTC"
                reference:
                  type: string
                  description: Reference ID from registration
                  example: "00042"
                requested_amount:
                  type: string
                  description: Amount user wants to send
                  example: "1.5"
      responses:
        '200':
          description: Amount suggestions calculated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      asset:
                        type: string
                        example: "BTC.BTC"
                      reference:
                        type: string
                        example: "00042"
                      reference_length:
                        type: integer
                        example: 5
                      decimals:
                        type: integer
                        example: 8
                      requested_amount:
                        type: string
                        example: "1.5"
                      normalized_amount:
                        type: string
                        description: Input amount normalized to asset decimals
                        example: "1.50000000"
                      suggestions:
                        type: object
                        properties:
                          round_up:
                            type: object
                            properties:
                              amount:
                                type: string
                                description: Suggested amount rounded up (always >= requested)
                                example: "1.50000042"
                              is_higher_than_requested:
                                type: boolean
                                example: true
                              description:
                                type: string
                                example: "Rounded up to ensure amount is higher than requested"
                          round_down:
                            type: object
                            properties:
                              amount:
                                type: string
                                description: Suggested amount rounded down (may be < requested)
                                example: "1.40000042"
                              is_lower_than_requested:
                                type: boolean
                                example: true
                              description:
                                type: string
                                example: "Rounded down to stay within requested limit"
                      minimum_amount_to_send:
                        type: string
                        description: Minimum valid amount for this reference ID
                        example: "0.00100042"
                      both_valid:
                        type: boolean
                        description: Whether both suggestions are above minimum amount
                        example: true
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Calculation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Asset:
      type: object
      properties:
        asset:
          type: string
          description: Asset identifier
          example: "BTC.BTC"
        decimals:
          type: integer
          description: Number of decimal places
          example: 8
        priceUSD:
          type: number
          format: float
          description: Current USD price
          example: 45000.50
        balanceRune:
          type: string
          description: RUNE balance in the pool
          example: "32134415450828"
        status:
          type: string
          description: Pool status
          example: "Available"
        isToken:
          type: boolean
          description: Whether this is a token (has contract address)
          example: false
    RegistrationRequest:
      type: object
      required:
        - asset
        - memo
      properties:
        asset:
          type: string
          description: |
            Asset to register memo for. **Must be from the list returned by GET /api/v1/assets**.
            Format: CHAIN.SYMBOL (e.g., BTC.BTC, ETH.ETH, GAIA.ATOM)
          pattern: '^[A-Z]+\.[A-Z0-9]+$'
          default: "ETH.ETH"
          example: "BTC.BTC"
        memo:
          type: string
          description: |
            THORChain transaction memo. Any non-empty string is accepted - THORChain will validate the format after registration.
            
            Common formats include:
            - `=:` for swaps (e.g., `=:ETH.ETH:0x742d35Cc6634C0532925a3b8D121C2c6C9b8DC8E:0/1/0`)
            - `+:` for adding liquidity (e.g., `+:BTC.BTC::bc1quser123:100`)
            - `-:` for withdrawing liquidity
            - `SWAP:`, `ADD:`, `WITHDRAW:` (alternative formats)
          minLength: 1
          default: "=:BTC.BTC:bc1quser123456789abcdef:0/1/0"
          example: "=:ETH.ETH:0x742d35Cc6634C0532925a3b8D121C2c6C9b8DC8E:0/1/0"
        requested_in_asset_amount:
          type: string
          description: |
            Optional. Amount requested by user for calculating a suggested amount that embeds the reference ID.
            When provided, the response will include a `suggested_in_asset_amount` that:
            - Embeds the reference ID in the decimal places according to the asset's decimal precision
            - Ensures the suggested amount is always higher than the requested amount
            - Handles minimum amount validation and decimal normalization
          example: "1.5"
    RegistrationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        internal_api_id:
          type: string
          description: Internal API tracking ID for this registration (unique identifier used by this API)
          example: "reg_abc123"
        asset:
          type: string
          description: Registered asset
          example: "BTC.BTC"
        memo:
          type: string
          description: Registered memo
          example: "=:ETH.ETH:0x1c7b17362c84287bd1184447e6dfeaf920c31bbe:1000:0:100"
        reference:
          type: string
          description: Reference ID for amount validation
          example: "20002"
        reference_length:
          type: integer
          description: Number of characters in the reference ID
          example: 5
        height:
          type: string
          description: Block height when memo was registered
          example: "12345678"
        registration_hash:
          type: string
          description: THORChain registration hash
          example: "AF524818D42B63D25BBA0CCC4909F127CAA645C0F9CD07324F2824CC151A64C8"
        registered_by:
          type: string
          description: THORChain address that registered the memo
          example: "thor1zupk5lmc84r2dh738a9g3zscavannjy3nzplwt"
        txHash:
          type: string
          description: THORChain transaction hash
          example: "A1B2C3D4E5F67890ABCDEF1234567890ABCDEF12"
        decimals:
          type: integer
          description: Number of decimal places for this asset
          example: 8
        minimum_amount_to_send:
          type: string
          description: Minimum amount that can be sent with this reference ID (calculated based on memoless encoding rules)
          example: "0.00100010"
        suggested_in_asset_amount:
          type: string
          description: |
            Optional. Calculated suggested amount that embeds the reference ID and is guaranteed to be higher than the requested amount.
            Only present when `requested_in_asset_amount` was provided in the request.
            
            The calculation logic:
            - If requested amount is below minimum_amount_to_send, returns the minimum
            - Otherwise, normalizes the amount to asset decimals (truncating, not rounding)
            - Embeds the reference ID in the last decimal places
            - Increments if needed to ensure suggested > requested
          example: "1.50020002"
    RegistrationStatus:
      type: object
      properties:
        registrationId:
          type: string
          example: "reg_abc123"
        status:
          type: string
          enum: [pending, confirmed, failed]
          example: "confirmed"
        txHash:
          type: string
          example: "A1B2C3D4E5F67890ABCDEF1234567890ABCDEF12"
        asset:
          type: string
          example: "BTC.BTC"
        memo:
          type: string
          example: "=:ETH.ETH:0x742d35Cc..."
        createdAt:
          type: string
          format: date-time
        referenceId:
          type: string
          description: Reference ID for amount validation (only when confirmed)
          example: "00123"
        depositInfo:
          $ref: '#/components/schemas/DepositInfo'
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: "INVALID_ASSET"
            message:
              type: string
              description: Human-readable error message
              example: "Asset BTC.INVALID is not supported"
            details:
              type: string
              description: Additional error details (development only)
            timestamp:
              type: string
              format: date-time
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional)

# Optional security requirement (uncomment if you implement API keys)
# security:
#   - ApiKeyAuth: []